link_storage: decouple AutoclosingSqliteConnection from connection creation
AutoclosingSqliteConnection should be concerned with implementing
a context manager providing an sqlite connection which closes
on __exit__. The fact that such connection is on-disk is another matter,
and it might as well not be the case.

This extracts the connection provider as an explicit dependency of
that class, enabling better testing of it.