Support saving links with more than one tag
The link -> tags relationship is one-to-many and is implemented using a separate
table holding the tags.

What this commit does:

* Create a new class to handle interaction with that table.

* Enhance the SQL creating tables (enforce not null and uniqueness of URLs).

* Refactor SqliteConnectionFactory in order to enforce foreign
key constraints on all sqlite connections, and allow the factory
to be used by ProviderOfConnectionToOnDiskSqliteDatabase.

* Refactor ApplicationDataDirectory to only use `@property`s, making
the class declaration rather cleaner.

* Remove the InMemorySqliteStorage: it was useful for bootstrapping
and defining the interface of the link storage, but it is useless now:
persistence is needed and the in-memory mode of sqlite works great for
testing. There's little point in maintaining it or further developing
it.