Fix ignoring mamba names when running flake8
The behaviour is exactly the same as before: we still run the risk
of ignoring some of those names in non-spec files, where it could be
dangerous. That is an unlikely scenario, however, given those names
are rather specific to spec files (save for `self`, which is itself
almost exclusive to methods, where it is always defined, and an error
would likely be raised otherwise).