linkstore: remove unit tests
They are not really testing anything valuable. They are all about checking the
delegation to the link storage, so no more than a bland repetition of the
implementation.

We could add tests checking that the query methods return Links and not
LinkRecords, but since Python is dynamic and LinkRecord is, for now,
substitutable for Link objects, that too would add close to zero value.