Start test-driving a links repository
We TDD the in-memory repo with the intention of:
	- discovering exactly what we want out of a links repositoroy in this system
	- develop a suite of tests to later apply to a persistent (sqlite) repository

As we go, we integrate this repo in parallel into the links service, adapting
the unit tests to use it instead of the in-memory sqlite storage (the current
implementation).

These tests are not intended to be thorough, just that the basic wiring is set
up. Thoroughness of e.g. filtering is tested at the unit level.