Remove link_id & LinkRecord kludge by introducing a Handles object
Handles is midway between a service and a repository; it contains little logic
and is mostly a pretty interface to a table. I decided to not give it the status
of a service for now because it's a UI-specific requirement. For the same
reason, handles are not included in the links table or, for that matter, even
known about in the whole links service.